Replace/Fail if `REPLACE_EXISTING` is specified (Windows)

+++ src/main/java/de/borisskert/boxfs/windows/BoxFsFileSystemProvider.java
@@ -132,6 +132,10 @@ class BoxFsFileSystemProvider extends FileSystemProvider {
         if (node.get().attributes().isDirectory()) {
             if (fileTree.exists(target)) {
                 if (replaceExisting) {
+                    if (fileTree.isDirectory(target) && hasChildren(target)) {
+                        throw new DirectoryNotEmptyException(target.toString());
+                    }
+
                     deleteRecursively(target);
                 } else {
                     throw new FileAlreadyExistsException(target.toString());
@@ -150,6 +154,12 @@ class BoxFsFileSystemProvider extends FileSystemProvider {
         }
 
         delete(source);
+    }
+
+    private boolean hasChildren(Path path) throws IOException {
+        try (DirectoryStream<Path> stream = newDirectoryStream(path, entry -> true)) {
+            return stream.iterator().hasNext();
+        }
     }
 
     private void deleteRecursively(Path path) throws IOException {
